Rewrite backend to handle the PipelineSchedule model
Now the model changed name, and fields are all on one model, some
backend changes can now be made to reflect that. Most notable changes
are the worker, which is renamed to PipelineSchedular worker, and the
PipelineSchedulesFinder.

The last change makes a tiny class right now, but is most probably to be
expended later.